Move publishEvent to RabbitMQ.ts
